add moshi dependencies
