+++ app/build.gradle
@@ -27,6 +27,7 @@ dependencies {
     // latest versions
     def koin_version = '2.2.3'
     def retrofit_version = "2.9.0"
+    def moshi_version = "2.9.0"
 
 
     // modules
@@ -45,7 +46,11 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
 
-    //retrofit
+    //retrofit and moshi
     implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
+    implementation 'com.squareup.retrofit2:adapter-rxjava3:2.9.0'
+    implementation "com.squareup.retrofit2:converter-moshi:$moshi_version"
+    implementation  ("com.squareup.moshi: moshi-kotlin:1.12.0")
+    //kapt "com.squareup.moshi:moshi-kotlin-codegen:1.12.0"
 
 }
