+++ build.gradle
@@ -3,6 +3,7 @@ buildscript {
     ext.kotlin_version = "1.4.32"
     repositories {
         google()
+        mavenCentral()
         jcenter()
     }
     dependencies {
@@ -17,6 +18,7 @@ buildscript {
 allprojects {
     repositories {
         google()
+        mavenCentral()
         jcenter()
     }
 }
