+++ core/network/build.gradle
@@ -6,13 +6,20 @@ dependencies {
     // latest versions
     //obs: verificar com o wesley se tem como colocar as versões em um lugar só
     def retrofit_version = "2.9.0"
+    def moshi_version = "2.9.0"
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.2'
     testImplementation 'junit:junit:4.+'
 
-    //retrofit
+    //retrofit and moshi
     implementation "com.squareup.retrofit2:retrofit:$retrofit_version"
+    implementation 'com.squareup.retrofit2:adapter-rxjava3:2.9.0'
+    implementation "com.squareup.retrofit2:converter-moshi:$moshi_version"
+    implementation  ("com.squareup.moshi: moshi-kotlin:1.12.0")
+    //kapt "com.squareup.moshi:moshi-kotlin-codegen:1.12.0"
+
+
 
 
 }
