--- core/network/src/main/java/com/shayanne/core/network/NetworkModule.kt
@@ -1,17 +1,0 @@
-package com.shayanne.core.network
-
-import okhttp3.OkHttpClient
-import retrofit2.Retrofit
-
-fun networkModule(
-    baseUrl: String,
-){
-    val okHttpClient = OkHttpClient.Builder()
-
-    Retrofit.Builder()
-        .baseUrl(baseUrl)
-        .client(okHttpClient.build())
-        .build()
-        .create(EmporiumService::class.java)
-
-}
