+++ core/network/src/main/java/com/shayanne/core/network/adapter/RepositoryAdapter.kt
@@ -1,0 +1,7 @@
+package com.shayanne.core.network.adapter
+
+import com.squareup.moshi.JsonAdapter
+
+class RepositoryAdapter : JsonAdapter.Factory {
+
+}
